pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shaket"
version = "0.1.0"
description = "An open protocol for multi-agent negotiation and auction"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Shaket Labs", email = "contact@shaket.dev"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "uvicorn>=0.27.0",
    "fastapi>=0.95.0",
    "litellm",
    "python-dotenv",
    "a2a-sdk[http-server]>=0.3.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["shaket*"]
